More improvements, refactored a bit
