+++ components/SendDataToggle.js
@@ -1,37 +1,15 @@
 import { Layout, Toggle } from "@ui-kitten/components";
-import dayjs from "dayjs";
+
 import * as Location from "expo-location";
 import * as SecureStore from "expo-secure-store";
 import * as TaskManager from "expo-task-manager";
 import React from "react";
 import { StyleSheet } from "react-native";
 import showToast from "../util/showToast";
-import { db } from "../stores/database";
+import insertLocation from "../util/insertLocation";
+import env from "../env";
 
 const LOCATION_TASK_NAME = "background-location-task";
-const FOREGROUND_LOCATION_SECONDS = 30;
-
-// Function to print and insert the location in the db
-const insertLocation = (locationObject) => {
-  const latitude = locationObject?.coords.latitude;
-  const longitude = locationObject?.coords.longitude;
-  const timestamp = dayjs(locationObject?.timestamp)
-    .format("YYYY-MM-DD HH:mm:ss.SSS")
-    .toString();
-
-  console.log(
-    `latitude:${latitude} longitude:${longitude} timestamp:${timestamp}`
-  );
-  db.transaction((tx) => {
-    tx.executeSql(
-      `
-            insert into coordinates (timestamp, latitude, longitude)
-            values (?, ?, ?)
-            `,
-      [timestamp, latitude, longitude]
-    );
-  });
-};
 
 // Apparently, it runs in the background 'a few times' each hour
 TaskManager.defineTask(LOCATION_TASK_NAME, ({ data: { locations }, error }) => {
@@ -66,10 +44,8 @@ export default function SendDataToggle() {
       try {
         const locationObject = await Location.getCurrentPositionAsync();
         insertLocation(locationObject);
-      } catch (error) {
-        console.log(error.message);
-      }
-    }, FOREGROUND_LOCATION_SECONDS * 1000);
+      } catch (error) {}
+    }, env.FOREGROUND_LOCATION_SECONDS * 1000);
 
     return () => clearInterval(foregroundLocationInterval);
   }, []);
