+++ env.js
@@ -1,5 +1,11 @@
 // Setting env variables in expo is finnicky, might as well use this
 
 const BACKEND_URL = "http://192.168.0.105:8000";
+const UPDATE_REMOTE_DB_INTERVAL_SECONDS = 60;
+const FOREGROUND_LOCATION_SECONDS = 30;
 
-export default { BACKEND_URL };
+export default {
+  BACKEND_URL,
+  UPDATE_REMOTE_DB_INTERVAL_SECONDS,
+  FOREGROUND_LOCATION_SECONDS,
+};
