+++ util/insertLocation.js
@@ -1,0 +1,26 @@
+import { db } from "../stores/database";
+import dayjs from "dayjs";
+
+// Function to print and insert the location in the db
+const insertLocation = (locationObject) => {
+  const latitude = locationObject?.coords.latitude;
+  const longitude = locationObject?.coords.longitude;
+  const timestamp = dayjs(locationObject?.timestamp)
+    .format("YYYY-MM-DD HH:mm:ss.SSS")
+    .toString();
+
+  console.log(
+    `latitude:${latitude} longitude:${longitude} timestamp:${timestamp}`
+  );
+  db.transaction((tx) => {
+    tx.executeSql(
+      `
+              insert into coordinates (timestamp, latitude, longitude)
+              values (?, ?, ?)
+              `,
+      [timestamp, latitude, longitude]
+    );
+  });
+};
+
+export default insertLocation;
